refactor: delete globalfuncs, globalvars, console; relocate live symbols
globalfuncs.py: 18 of 19 functions were dead code superseded by the 2.0
orchestration layer (handler registries, Config model, rich.Live progress,
ContentHandler chunking). Sole survivor is_admin() inlined into run.py as
_is_admin(); the lazy-import workaround for a circular dependency is no
longer needed.

globalvars.py: ERROR_CODES was the only dead entry (both callers deleted in
Phase 4). The four live constants (MAX_CHUNK_SIZE, DEFAULT_CHUNK_COUNT,
EXCEL_BLANK_ROW_LIMIT, EXCEL_BLANK_COL_LIMIT) moved to
filehandlers/_constants.py where they belong — nothing outside filehandlers
ever imported them.

console.py: sole caller was globalfuncs.progress_line(), itself dead code.
Deleted with no remaining callers.

tests/test_make_chunks.py and tests/test_write_default_config.py deleted —
both tested dead globalfuncs functions with no production callers.

All 259 tests pass; ruff and mypy clean.
