+++ src/piidigger/filehandlers/_constants.py
@@ -1,0 +1,4 @@
+MAX_CHUNK_SIZE = 650
+DEFAULT_CHUNK_COUNT = 100_000
+EXCEL_BLANK_ROW_LIMIT = 250
+EXCEL_BLANK_COL_LIMIT = 500
