+++ src/piidigger/filehandlers/plaintext.py
@@ -1,8 +1,8 @@
 from collections.abc import Iterator
 
+from piidigger.filehandlers._constants import DEFAULT_CHUNK_COUNT, MAX_CHUNK_SIZE
 from piidigger.filehandlers._sharedfuncs import ContentHandler
 from piidigger.getencoding import detect_encoding
-from piidigger.globalvars import DEFAULT_CHUNK_COUNT, MAX_CHUNK_SIZE
 
 HANDLES = {
     "ext": [
