--- src/piidigger/globalvars.py
@@ -1,11 +1,0 @@
-ERROR_CODES = {
-    "ok": 0,
-    "userTerm": 1,
-    "invalidConfig": 1,
-    "unknownError": 1024,
-}
-
-DEFAULT_CHUNK_COUNT = 100_000
-EXCEL_BLANK_ROW_LIMIT = 250
-EXCEL_BLANK_COL_LIMIT = 500
-MAX_CHUNK_SIZE = 650
